requirements.txt
bcrypt==3.1.4
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
facebook-sdk==2.0.0
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
peewee==3.0.11
pkg-resources==0.0.0
psycopg2==2.7.3.2
pycparser==2.18
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
